Add event handling
This change introduces a new way to pass data between components.  Now,
instead of having AddFeedform tell FeedPage that the user wants to add a
new feed (so that FeedList can add the feed and update the list), this
fires off an event that FeedList subscribes to.

Of course, in this particular case we could just have FeedList have a
plain old event listener for the button in the form, but I figure
there'll be more issues like this in the future, so it might be
appropriate to have a custom event system.