Move most state handling to FeedPage, and get rid of routing
I moved all of the stateful logic (including GET and POST requests) to
FeedPage. This makes the code more idiomatic React, and hopefully will
also make things easier to test and reason about.

En route to that I got rid of the routing stuff. It was a cool idea, but
it was making the code too complicated at such an early
stage. Crucially, react-router has its own opinions about how to pass
props to children, and this was making it difficult to make compatible
with the existing code.